bunch of small code tweaks. Making sure everything plays well with the Postman tests that I'm re-using from other projects. working.

+++ request_handler_a5.js
@@ -303,6 +303,37 @@ const ALL_BOAT_ATTRIBUTES = ['name', 'length', 'type', 'id'];
         return;
     }
 
+    /*
+    // does boat resource exist?
+    var boat = await db.getBoat(req.params.id);
+    console.log('Boat to be deleted is');
+    console.log(boat);
+    if (utils.isEmpty(boat)) {
+        res.status(404).send({"Error": "No boat with this boat_id exists"});
+        return;
+    }
+    */
+
+    // update all loads on boat
+    var loads = await db.getLoadsFromBoat(boat, req);
+    for (i=0; i<loads.length; i++) {
+        loads[i].carrier = null;
+        let success = await db.updateLoad(JSON.parse(JSON.stringify(loads[i])));
+        console.log('load updated: ' + success);
+    }
+
+    /*
+    // delete boat resource
+    var deleted = await db.deleteResource(db.BOAT, boat);
+    if (!deleted) {
+        console.log('ERROR: Boat was not deleted');
+        res.status(500).send({"Error": "Internal database error"});
+        return;
+    }
+
+    res.status(204).send();
+    */
+
     // delete
     var deleted = await db.deleteResource('Boats', boat);
     if (!deleted) {
